censorship-scenarios/HTTP-Blocking/requirements.txt
# HTTP-Blocking Scenario Dependencies

# Core dependencies
scapy>=2.4.5
docker>=6.0.0
paramiko>=3.0.0
PyYAML>=6.0

# OpenEvolve integration
openevolve>=0.1.0

# HTTP utilities
requests>=2.28.0
urllib3>=1.26.0
h2>=4.0.0

